SDMMC performance improvements
Move LwIP DMA buffers to end of D1 (AXI SRAM) memory (they need to
be in D1 RAM to support DMA access from the SDMMC1 peripheral).

Allocate a 64KB section in the linker file, and use this for the
LwIP DMA buffers and descriptors. (Need to better understand the
actual buffer usage to size this more appropriately).

Note that absolute addresses are needed for the MPU configuration
to support L1 cache operations, and buffers are aligned on 32byte
boundaries to match L1 cache lines.

Configure LwIP to use 32byte alignment for dynamic allocations when
L1 cache is enabled, increase the size of the TX heap, and disable
the 2*MSS restriction on send buffer.