USB HS peripheral support
The STM32H723 part does not have an on-chip USB FS interface, so we
need to use the USB HS interface in FS mode..

This requires different STM32CubeIDE support files, which have been
put into a new folder. IDE configurations will be updated to use
this new folder for H723 builds only.

Note that the existing H743 builds can not use the HS interface
in FS mode, as the two interfaces are tied to different pins.

Note: Add call to usbBufferInput().