Update driver.json
I see usb_cdc is set to 0 for the Nucleo boards, the F756 I have support both USB <> UART via the ST-link interface and native USB and the F7 does as well?

New is is number of  i2c_ports, I have set this to 1 at the driver level, if a board has a connector for I2C it should be set at the board level as well.